pyproject.toml
[project]
name = "ai-training-for-devs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastembed>=0.7.4",
    "fastmcp>=2.14.0",
    "gradio>=6.1.0",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langfuse>=3.10.5",
    "langgraph>=1.0.4",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "ruff>=0.14.8",
]
